[RSAENH] * Sync with Wine 1.7.1. CORE-7469
svn path=/trunk/; revision=60373

+++ reactos/dll/win32/rsaenh/rsaenh.c
@@ -362,7 +362,7 @@ static struct handle_table handle_table;
  *
  * Initializes and destroys the handle table for the CSP's handles.
  */
-int WINAPI DllMain(HINSTANCE hInstance, DWORD fdwReason, PVOID pvReserved)
+BOOL WINAPI DllMain(HINSTANCE hInstance, DWORD fdwReason, PVOID reserved)
 {
     switch (fdwReason)
     {
@@ -373,10 +373,11 @@ int WINAPI DllMain(HINSTANCE hInstance, DWORD fdwReason, PVOID pvReserved)
             break;
 
         case DLL_PROCESS_DETACH:
+            if (reserved) break;
             destroy_handle_table(&handle_table);
             break;
     }
-    return 1;
+    return TRUE;
 }
 
 /******************************************************************************
@@ -1701,6 +1702,11 @@ static BOOL unpad_data(CONST BYTE *abData, DWORD dwDataLen, BYTE *abBuffer, DWOR
 {
     DWORD i;
     
+    if (dwDataLen < 3)
+    {
+        SetLastError(NTE_BAD_DATA);
+        return FALSE;
+    }
     for (i=2; i<dwDataLen; i++)
         if (!abData[i])
             break;
@@ -2378,11 +2384,13 @@ BOOL WINAPI RSAENH_CPDecrypt(HCRYPTPROV hProv, HCRYPTKEY hKey, HCRYPTHASH hHash,
                     *pdwDataLen -= pbData[*pdwDataLen-1];
                 else {
                     SetLastError(NTE_BAD_DATA);
+                    setup_key(pCryptKey);
                     return FALSE;
                 }
             }
             else {
                 SetLastError(NTE_BAD_DATA);
+                setup_key(pCryptKey);
                 return FALSE;
             }
         }
